Refactorizar la estructura de la aplicación añadiendo el componente Home, integrando Vitest para pruebas, y eliminando componentes y estilos no utilizados
