+++ src/App.jsx
@@ -1,22 +1,23 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom';
-import Login from './components/Login1';     // Asegúrate que esté bien el nombre del archivo
-import Register from './components/Register'; // Ruta correcta al componente Register
-import Recuperar from './components/Recuperar';
-import VerificarCodigo from './components/VerificarCodigo';
-import Restablecer from './components/Restablecer';
+import Login from './components/auth/login/Login1';     // Asegúrate que esté bien el nombre del archivo
+import Register from './components/auth/register/Register'; // Ruta correcta al componente Register
+import Recuperar from './components/auth/restore/Recuperar'; // Ruta correcta al componente Recuperar
+import VerificarCodigo from './components/auth/restore/VerificarCodigo'; // Ruta correcta al componente VerificarCodigo
+import Restablecer from './components/auth/restore/Restablecer'; // Ruta correcta al componente Restablecer
+import Home from './pages/app/home'; 
 
 function App() {
   return (
     <Router>
       <Routes>
-        <Route path="/" element={<Navigate to="/login" />} />
+        <Route path="/" element={<Navigate to="/home" />} />
         <Route path="/login" element={<Login />} />
         <Route path="/registro" element={<Register />} />
         <Route path="/recuperar" element={<Recuperar />} />
         <Route path="/verificar-codigo" element={<VerificarCodigo />} />
         <Route path="/restablecer" element={<Restablecer />} />
-        
+        <Route path="/home" element={<Home/> } />
       </Routes>
     </Router>
   );
