+++ src/components/auth/login/Login.css
@@ -1,16 +1,14 @@
-body, html, #root {
-  margin: 0;
-  padding: 0;
-  height: 100%;
-  background-color: #fff9ed;
-  font-family: 'Segoe UI', sans-serif;
-}
+
 
 .login-page {
   display: flex;
   flex-direction: column;
-  min-height: 100vh;
+  align-items: center;
+  justify-content: center;
+  height: 100vh;
+  width: 100vw;
 }
+
 .form-box h2 {
   color: rgba(130, 23, 23, 1); /* 👈 o el color que tú quieras */
 }
@@ -19,6 +17,8 @@ body, html, #root {
   display: flex;
   justify-content: center;
   align-items: center;
+  width: 100%;
+  height: 100%;
 }
 
 .form-box {
