+++ src/components/auth/register/Register.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import axios from 'axios';
-import '/src/styles/Register.css';
+import './Register.css';
 
 const Register = () => {
   const [form, setForm] = useState({
