+++ src/main.test.jsx
@@ -1,0 +1,7 @@
+import { describe, it, expect } from 'vitest';
+
+describe('Main Application Logic', () => {
+    it('hello world!', () => {
+        expect(1 + 1).toBe(2);
+    });
+});
