+++ src/pages/app/home.css
@@ -1,0 +1,62 @@
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+  };
+  
+  :root {
+    --color-primary: #111828;           /* Negro ligeramente más suave */
+    --color-secondary: #1E1E1E;        /* Azul oscuro profesional */
+    --color-tertiary: #242424;         /* Azul-gris para profundidad */
+    --color-quaternary: #EDEDED;        /* Gris muy claro para fondos */
+    --color-quinary: #F5F5F5;           /* Blanco roto para contraste */
+    --color-accent: #3D84B8;           /* Azul corporativo para acentos */
+    --color-success: #B1F6BB;          /* Verde para acciones positivas */
+    --color-txtsuccess: #007712;             /* Verde para acciones positivas */
+    --color-warning: #ECFE48;          /* Amarillo para advertencias */
+    --color-txtwarning: #776D00;
+    --color-danger: #F6B1B1;           /* Rojo para acciones críticas */
+    --color-text: #333333;             /* Texto principal más suave */
+    --color-septenary: #FFFFFF;        /* Blanco puro */
+    --color-txtdanger: #821717;
+  
+  }
+  
+  
+  .container {
+    display: grid;
+    grid-template-columns: repeat(10,1fr);
+    grid-template-rows: repeat(10,1fr);
+    gap: 0px;
+    height: 100vh;
+    width: 100vw;
+    background-color: var(--color-secundary);
+    overflow-wrap: normal;
+    overflow: hidden;
+  }
+  
+  .navbar-container {
+    display: grid;
+    grid-column: span 10 / span 10;
+    grid-row: span 1 / span 1;
+    background-color: var(--color-primary);
+  }
+  
+  .sidebar-container {
+    display: grid;
+    grid-column: span 2 / span 2;
+    grid-row: span 10 / span 10;
+    background-color: var(--color-secundary);
+  }
+  
+  .content-container {
+    display: grid;
+    grid-column: span 8 / span 8;
+    grid-row: span 10 / span 10;
+    background-color: var(--color-tertiary);
+  }
+  
+  .en-proceso{
+      background-color: var(--color-warning);
+      color: var(--color-text);
+  }
