+++ src/pages/app/home.jsx
@@ -1,0 +1,25 @@
+import { useState } from 'react';
+import './home.css';
+import Navbar from '../../components/navbar/nabvar';
+import Sidebar from '../../components/sidebar/sidebar';
+import Graficas from '../graficas/graficas';
+
+const Home = () => {
+    return(
+        <>
+        <main className='container'>
+          <nav className='navbar-container'>
+            <Navbar/>    
+          </nav>
+          <aside className='sidebar-container'>
+            <Sidebar/>
+          </aside>
+          <div className='content-container'>
+            <Graficas/>
+          </div>
+        </main>
+        </>
+    )
+}
+
+export default Home;
